Don't map in Select mode if appropriate
textobj#user#plugin() defines key mappings to select a text object in
Visual mode and Select mode.  But key mappinpgs which start with
a printable character should not be defined in Select mode, because
users will expect printable characters to replace the current selection
in Select mode (according to :help Select-mode-mappings).

This problem is rediscovered by the recent trend of snippet plugins
which utilize the behavior of Select mode.