Keep count for "select-x-function"
s:prepare_selection, more precisely the gv<Esc> hack, was introduced by
the commit c12d4ca719aaef7ae59642676f6c81ad81a47af2.  The hack might be
necessary for old versions of Vim, but it seems to be verbose now.  We
can cancel the current operation and and restore the last selection
without the hack.

And more importantly, the hack resets a count.  So that it is not
possible to know a count given to a text object typed in Visual mode
within "select-x-function" via v:count and v:count1.

Therefore, s:prepare_selection should be removed.