Ensure the matchlist is sufficiently long before accessing it.
When running Vim in verbose mode, the `silent scriptnames` line in
`s:snr_prefix` will cause the result to contain a line such as the
following at the beginning of the result:
    ^@line 2:   silent scriptnames

This line does not match the pattern, so an empty list is returned.
However, the subsequent line attempts to access the empty list causing
an error.  Let's check the list before accessing the elements.