Restore cursor position in s:cancel_selection always
I've noticed that `vif` (using
[vim-textobj-python](https://github.com/bps/vim-textobj-python.git))
would cause the window to scroll when the `normal! gv` was called.
Restoring the cursor position also for visual mode fixed this for me.

Apparently it breaks no test, but there should probably be one added for
this - in case it's not caused by some other plugin / config from me.